requirements.txt
tweepy==4.14.0
python-telegram-bot==20.3
apscheduler==3.10.1
python-dotenv==0.21.0
finnhub
requests==2.31.0